:ok_hand: F06 > NavBar > add Toast component + refacto
